requirements.txt
requests
slacker
python-telegram-bot